Add exceptions to services (#135)
* added exception to DeleteFood method

* Added exception to GetRestaurantOfFood method

* Added exception to GetTypeOfFood method

* Added exeption to ChangePass method

* Added exception to GetRestaurantDtoFromMail method

* Added exception to DeleteAccount method

* Added exception to GetRestaurantById method

* Changed exception messages

* Changed naming for consistency in FoodService.cs

* Changed naming for consistency in RestaurantService.cs

+++ backend/Services/Services/RestaurantService.cs
@@ -17,25 +17,50 @@ namespace Services.Services
         {
             _restaurantRepository = restaurantRepository;
         }
-
-        public void ChangePass(Mail email, Password newPassword)
+        
+        public void ChangePass(Mail mail, Password newPassword)
         {
-            Restaurant restaurant = _restaurantRepository.GetByMail(email);
-            Credentials creds = restaurant.Credentials;
-            creds.Password = newPassword;
-
-            _restaurantRepository.Update(restaurant);
+            Restaurant restaurant = _restaurantRepository.GetByMail(mail);
+            if(restaurant == null)
+            {
+                throw new System.Exception("Invalid email.");
+            }
+            else
+            {
+                Credentials creds = restaurant.Credentials;
+                creds.Password = newPassword;
+                _restaurantRepository.Update(restaurant);
+            }
+            
+            
         }
         
         public RestaurantDto GetRestaurantDtoFromMail(Mail mail)
         {
-            return RestaurantDto.FromEntity(_restaurantRepository.GetByMail(mail));
+            Restaurant restaurant = _restaurantRepository.GetByMail(mail);
+            if(restaurant == null)
+            {
+                throw new System.Exception("Invalid email.");
+            }
+            else
+            {
+                return RestaurantDto.FromEntity(restaurant);
+            }
+            
         }
-
+        
         public void DeleteAccount(Credentials creds)
         {
             Restaurant restaurant = _restaurantRepository.GetByMail(creds.Mail);
-            _restaurantRepository.Delete(restaurant.Id);
+            if (restaurant == null)
+            {
+                throw new System.Exception("Invalid credentials.");
+            }
+            else
+            {
+                _restaurantRepository.Delete(restaurant.Id);
+            }
+            
         }
 
         public IEnumerable<RestaurantDto> GetAllRestaurants()
@@ -44,11 +69,19 @@ namespace Services.Services
                     .GetAll()
                     .Select(r => RestaurantDto.FromEntity(r));
         }
-
-        public RestaurantDto GetRestaurantById(string id)
+        
+        public RestaurantDto GetRestaurantById(string idRestaurant)
         {
-            Restaurant restaurant = _restaurantRepository.GetById(id);
-            return RestaurantDto.FromEntity(restaurant);
+            Restaurant restaurant = _restaurantRepository.GetById(idRestaurant);
+            if (restaurant == null)
+            {
+                throw new System.Exception("Invalid id.");
+            }
+            else
+            {
+                return RestaurantDto.FromEntity(restaurant);
+            }
+            
         }
 
         public IEnumerable<RestaurantDto> GetRestaurantsNear(Coords coords)
@@ -82,7 +115,7 @@ namespace Services.Services
             }
             else
             {
-                throw new System.Exception("• There is already an account registered on this mail");
+                throw new System.Exception("There is already an account registered on this mail.");
             }
         }
